fix: 32-bit platforms panicked on unaligned 64-bit atomics
On 32-bit GOARCH (arm/386, e.g. Android TV boxes) a bare uint64 inside a
struct is only 4-byte aligned, so every sync/atomic op on it faults with
panicUnaligned. ServerSession's sendSeq/sendPacketNo/recvSeq and the
Server health counters sat at 4-byte offsets and would crash the server on
first DATA record / first decode failure; clientSession happened to be
aligned by luck.

- Convert all 64-bit fields accessed via sync/atomic to atomic.Uint64
  (compiler-aligned to 8 bytes on every GOARCH via align64).
- Add tunnel/alignment.go compile-time assertions: any field that drifts
  off an 8-byte offset fails the build on every platform.
- Fix a pre-existing 32-bit int overflow (4294967295) in the ping-group
  range test.

Validated: build + vet + gofmt + go test -count=1, plus GOOS=linux
GOARCH=arm/386 build and vet.

+++ tunnel/icmp_ping_test.go
@@ -333,7 +333,7 @@ func TestPingPumpEndsWhenTheConnCloses(t *testing.T) {
 func TestParsePingGroupRange(t *testing.T) {
 	cases := []struct {
 		in          string
-		min, max    int
+		min, max    int64 // int64 so 4294967295 fits on 32-bit platforms too
 		ok          bool
 		description string
 	}{
@@ -349,7 +349,7 @@ func TestParsePingGroupRange(t *testing.T) {
 	for _, tc := range cases {
 		t.Run(tc.description, func(t *testing.T) {
 			min, max, ok := parsePingGroupRange(tc.in)
-			if ok != tc.ok || (ok && (min != tc.min || max != tc.max)) {
+			if ok != tc.ok || (ok && (int64(min) != tc.min || int64(max) != tc.max)) {
 				t.Fatalf("parsePingGroupRange(%q) = %d, %d, %t; want %d, %d, %t",
 					tc.in, min, max, ok, tc.min, tc.max, tc.ok)
 			}
